fix path in github mode

+++ seismiqb/tests/geometry_test.py
@@ -5,6 +5,7 @@ import os
 import pytest
 import pprint
 from datetime import date
+
 from ..batchflow.utils_notebook import run_notebook
 
 
@@ -13,46 +14,69 @@ DATESTAMP = date.today().strftime("%Y-%m-%d")
 DROP_EXTRA_FILES = True
 SHOW_TEST_ERROR_INFO = True
 TESTS_SCRIPTS_DIR = os.getenv("TESTS_SCRIPTS_DIR", os.path.dirname(os.path.realpath(__file__))+'/')
-TEST_FOLDER = os.path.join(TESTS_SCRIPTS_DIR, 'notebooks/geometry_test_files/')
+TEST_DIR = os.path.join(TESTS_SCRIPTS_DIR, 'notebooks/geometry_test_files/')
 SHOW_MESSAGE = True
+GITHUB_MODE = True
 
-@pytest.fixture(scope="session")
-def tests_notebook(tmpdir_factory):
+def test_geometry(capsys, tmpdir):
     """ Run SeismicGeometry test notebook."""
     # Delete old test notebook results
-    previous_output_files = glob.glob(os.path.join(TESTS_SCRIPTS_DIR, 'notebooks/geometry_test_out_*.ipynb'))
+    if GITHUB_MODE:
+        SAVING_DIR = tmpdir.mkdir("notebooks").mkdir("geometry_test_files")
+        out_path_ipynb = SAVING_DIR.join(f"geometry_test_out_{DATESTAMP}.ipynb")
 
-    for file in previous_output_files:
-        os.remove(file)
+    else:
+        # Clear outdatted files
+        previous_output_files = glob.glob(os.path.join(TESTS_SCRIPTS_DIR, 'notebooks/geometry_test_out_*.ipynb'))
 
-    out_path_ipynb = tmpdir_factory.mktemp("notebooks") / "geometry_test_out_{DATESTAMP}.ipynb"
+        for file in previous_output_files:
+            os.remove(file)
+
+        # Path to a new test noteboook result
+        SAVING_DIR = TEST_DIR
+        out_path_ipynb = os.path.join(TESTS_SCRIPTS_DIR, f'notebooks/geometry_test_out_{DATESTAMP}.ipynb')
 
     # Tests execution
-    _, tests_notebook = run_notebook(
+    exec_info, tests_notebook = run_notebook(
         path=os.path.join(TESTS_SCRIPTS_DIR, 'notebooks/geometry_test.ipynb'),
         nb_kwargs={
-            'TEST_FOLDER': TEST_FOLDER,
-            'NOTEBOOKS_FOLDER': os.path.join(TESTS_SCRIPTS_DIR, 'notebooks/'),
+            'NOTEBOOKS_DIR': os.path.join(TESTS_SCRIPTS_DIR, 'notebooks/'),
+            'TEST_DIR': TEST_DIR,
             'DATESTAMP': DATESTAMP,
             'DROP_EXTRA_FILES': DROP_EXTRA_FILES,
-            'SHOW_TEST_ERROR_INFO': SHOW_TEST_ERROR_INFO
+            'SHOW_TEST_ERROR_INFO': SHOW_TEST_ERROR_INFO,
+            'SAVING_DIR': SAVING_DIR,
+            'GITHUB_MODE': GITHUB_MODE
         },
-        insert_pos=0,
+        insert_pos=1,
         out_path_ipynb=out_path_ipynb,
         display_links=False,
         return_nb=True
     )
-    return tests_notebook
 
-def test_geometry(capsys, tests_notebook):
-    # Get message and timing from the notebook cells (github don't save temporary files)
-    print(tests_notebook)
-    msg = tests_notebook['cells'][7]['outputs'][0]['text']
-    timings = json.loads(tests_notebook['cells'][9]['outputs'][0]['text'].replace('\'', '\"'))
+    if exec_info is True:
+        # Open message
+        message_path = glob.glob(os.path.join(SAVING_DIR, 'message_*.txt'))[-1]
+
+        with open(message_path, "r") as infile:
+            msg = infile.readlines()
+
+        # Open timings
+        timings_path = glob.glob(os.path.join(SAVING_DIR, 'timings_*.json'))[-1]
+
+        with open(timings_path, "r") as infile:
+            timings = json.load(infile)
+
+    else:
+        msg = ['SeismicGeometry tests execution failed.\n']
+        timings= {'state': 'FAIL'}
 
     with capsys.disabled():
+        # Tests output
+
         if SHOW_MESSAGE:
-            print(msg, '\n')
+            for line in msg:
+                print(line)
 
         pp = pprint.PrettyPrinter()
         pp.pprint(timings)
@@ -61,5 +85,4 @@ def test_geometry(capsys, tests_notebook):
         if timings['state']=='OK':
             print('Tests for SeismicGeometry were executed successfully.\n')
         else:
-            print(f'SeismicGeometry tests failed.\n')
-            assert False
+            assert False, 'SeismicGeometry tests failed.\n'
